fix: property detail 500, map race condition, listing prices, API diagnostic
- Fix MissingGreenlet 500 on property detail: use selectinload for
  price_history and tax_history relationships so they are eagerly fetched
  in the same async query instead of triggering lazy I/O after the fact
- Fix map marker race condition: track mapReady state via map 'load'
  event; marker effect now re-runs when the map becomes ready, so
  properties that arrived before the map finished initializing still
  render as dots
- Fix missing prices: switch RentCast search_properties from /properties
  (no listing price) to /listings/sale (includes asking price); falls
  back to /properties if the listings endpoint fails
- Fix Decimal→string serialization: change baths, lot_size_acres, and
  current_price in PropertyBase schema from Decimal to float so they
  serialize as JSON numbers instead of quoted strings
- Clear stale null-price DB cache; next search re-fetches from listings
- Add backend/scripts/test_apis.py: runs connectivity + data-shape checks
  against RentCast, FRED, Census, ATTOM, FBI Crime, API Ninjas,
  Anthropic, and Mapbox; skips gracefully for unconfigured keys

API status after this run:
  PASS  RentCast (listings/sale + properties fallback)
  PASS  FRED (30yr mortgage rate = 6.3%)
  PASS  Census (ACS5 median income)
  FAIL  ATTOM (404 on all endpoints — likely trial plan restriction)
  SKIP  FBI Crime (key not configured)
  PASS  API Ninjas (city lookup)
  SKIP  Anthropic (key not configured)
  PASS  Mapbox (token valid)

+++ backend/app/schemas/property.py
@@ -88,12 +88,12 @@ class PropertyBase(BaseModel):
     lat: Optional[float] = None
     lng: Optional[float] = None
     beds: Optional[int] = None
-    baths: Optional[Decimal] = None
+    baths: Optional[float] = None
     sqft: Optional[int] = None
-    lot_size_acres: Optional[Decimal] = None
+    lot_size_acres: Optional[float] = None
     year_built: Optional[int] = None
     property_type: Optional[str] = None
-    current_price: Optional[Decimal] = None
+    current_price: Optional[float] = None
     list_date: Optional[date] = None
     days_on_market: Optional[int] = None
     photo_urls: Optional[list] = None
